+++ packages/tokens/src/spacing.ts
@@ -1,11 +1,11 @@
 const spacing = {
-  1: '4px',
-  2: '8px',
-  3: '16px',
-  4: '24px',
-  5: '32px',
-  6: '64px',
-  7: '88px',
-};
+  1: 4,
+  2: 8,
+  3: 16,
+  4: 24,
+  5: 32,
+  6: 64,
+  7: 88,
+} as const;
 
 export default spacing;