+++ packages/typography/src/Typography.tsx
@@ -4,12 +4,13 @@ import { HTMLElementProps } from '@leafygreen-ui/lib';
 import { useBaseFontSize } from '@leafygreen-ui/leafygreen-provider';
 import { css, cx } from '@leafygreen-ui/emotion';
 import { uiColors } from '@leafygreen-ui/palette';
+import { fontFamilies } from '@leafygreen-ui/tokens';
 import { typeScale1, typeScale2 } from './styles';
 
 const sharedStyles = css`
   margin: unset;
-  font-family: Akzidenz, 'Helvetica Neue', Helvetica, Arial, sans-serif;
-  color: ${uiColors.gray.dark2};
+  font-family: ${fontFamilies.default};
+  color: ${uiColors.gray.dark3};
 `;
 
 const h1 = css`
@@ -42,6 +43,23 @@ function H2({ children, className, ...rest }: H2Props) {
     <h2 {...rest} className={cx(sharedStyles, h2, className)}>
       {children}
     </h2>
+  );
+}
+
+const h3 = css`
+  font-size: 24px;
+  line-height: 28px;
+  letter-spacing: 0px;
+  font-weight: medium;
+`;
+
+type H3Props = HTMLElementProps<'h3'>;
+
+function H3({ children, className, ...rest }: H3Props) {
+  return (
+    <h3 {...rest} className={cx(sharedStyles, h3, className)}>
+      {children}
+    </h3>
   );
 }
 
@@ -85,10 +103,11 @@ function Body({ children, className, weight = 'regular', ...rest }: BodyProps) {
 }
 
 const code = css`
-  font-family: 'Source Code Pro', monospace;
   background-color: ${uiColors.gray.light3};
   border: 1px solid ${uiColors.gray.light1};
   border-radius: 3px;
+  font-family: ${fontFamilies.code};
+  display: inline-block;
 `;
 
 const codeLink = css`
@@ -148,6 +167,26 @@ function InlineCode({ children, className, ...rest }: InlineCodeProps) {
   return renderedInlineCode();
 }
 
+const inlineKeyCode = css`
+  font-family: ${fontFamilies.code};
+  color: ${uiColors.gray.dark3};
+  border: 1px solid ${uiColors.gray.dark3};
+  border-radius: 3px;
+  padding-left: 4px;
+  padding-right: 4px;
+`;
+
+function InlineKeyCode({ children, className, ...rest }: InlineCodeProps) {
+  const size = useBaseFontSize();
+  const body = size === 16 ? typeScale2 : typeScale1;
+
+  return (
+    <code className={cx(inlineKeyCode, body, className)} {...rest}>
+      {children}
+    </code>
+  );
+}
+
 const disclaimer = css`
   display: block;
   font-size: 12px;
@@ -179,4 +218,14 @@ const Overline: ExtendableBox<{
   return <Box className={cx(sharedStyles, overline, className)} {...rest} />;
 };
 
-export { H1, H2, Subtitle, Body, InlineCode, Disclaimer, Overline };
+export {
+  H1,
+  H2,
+  H3,
+  Subtitle,
+  Body,
+  InlineCode,
+  InlineKeyCode,
+  Disclaimer,
+  Overline,
+};