+++ website/pages/_app.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import type { AppProps } from 'next/app';
 
 function MyApp({ Component, pageProps }: AppProps) {