+++ website/pages/index.tsx
@@ -1,5 +1,6 @@
+import React from 'react';
 import { Body } from '@leafygreen-ui/typography';
 
 export default function Home() {
-  return <Body>Coming Soon!</Body>;
+  return <Body>Hello, World!</Body>;
 }