+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @leafygreen-ui/tokens
 
+## 0.3.0
+
+### Minor Changes
+
+- f792966: Adds default and code fonts to Tokens package
+
+## 0.2.0
+
+### Minor Changes
+
+- 6ca8fe6: Updates spacing to numerical rather than string values and exports as const so that consumers can see the values
+
 ## 0.1.1
 
 ### Patch Changes